Backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
opencv-python==4.8.1.78
face-recognition==1.3.0
numpy==1.24.3
Pillow==10.0.1
python-dotenv==1.0.0
bcrypt==4.0.1
marshmallow==3.20.1
werkzeug==2.3.7
scikit-learn==1.3.0
# PostgreSQL dependencies
psycopg2-binary==2.9.7
# Vector database dependencies
chromadb==0.4.15
faiss-cpu==1.7.4
# Additional vector operations
sentence-transformers==2.2.2
# Database connection pooling
sqlalchemy-pool==1.0.1
# ArcFace dependencies for 512D embeddings
insightface==0.7.3
onnxruntime==1.16.3
onnx==1.15.0